refactoring

+++ thesis/code/taglessfinal/src/Export/CSV.hs
@@ -3,7 +3,7 @@ module Export.CSV where
 import System.IO
 import Text.Printf
 
-writeCSVFile :: String -> [(Double, (Double, Double, Double))] -> IO ()
+writeCSVFile :: String -> [(Double, (Int, Int, Int))] -> IO ()
 writeCSVFile fileName xs = do
   hdl <- openFile fileName WriteMode
 
@@ -12,10 +12,10 @@ writeCSVFile fileName xs = do
 
   hClose hdl
 
-sirDynamicToString :: (Double, (Double, Double, Double)) -> String
+sirDynamicToString :: (Double, (Int, Int, Int)) -> String
 sirDynamicToString (t, (s, i, r)) =
   printf "%.4f" t ++
-  "," ++ printf "%.4f" s ++
-  "," ++ printf "%.4f" i ++ 
-  "," ++ printf "%.4f" r ++
+  "," ++ printf "%d" s ++
+  "," ++ printf "%d" i ++ 
+  "," ++ printf "%d" r ++
   ";"
